Rename regionAlpha3Codes to mapViewSettings
