+++ src/assets/mapViewSettings.js
